requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.11
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-import-export==3.0.1
django-select2
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
gunicorn
whitenoise
